Install ruby linters
