+++ Gemfile
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+source 'https://rubygems.org'
+
+gem 'rubocop', '>= 1.0', '< 2.0'
+
+# gem "rails"
