+++ README.md
@@ -1,0 +1,36 @@
+## my Enumerable
+
+In this project, we will create a custom myEnumerable method in ruby.
+
+
+## Authors
+
+👤 Daniel Shdeed
+
+- GitHub: [https://github.com/Danieldotcomcoder)]
+- Twitter: [https://twitter.com/DannyDotcoder)]
+- LinkedIn: [https://www.linkedin.com/in/daniel-shdeed-832b03115/]
+
+
+
+👤 **Karel van Oordt**
+
+- GitHub: [@karelvanoordt](https://github.com/karelvanoordt)
+- Twitter: [@karelvanoordt](https://twitter.com/karelvanoordt)
+- LinkedIn: [karelvanoordt](https://linkedin.com/in/karelvanoordt)
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](../../issues/).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Hat tip to anyone whose code was used
+- Inspiration
+- etc
